requirements.txt
Flask==2.3.2
scikit-learn==1.3.0
google-generativeai==0.3.0
